Don't copy *data* files already in db during import
Since the data files are content addressable, we shouldn't need
to copy a new one in. Since they are the ones that get mmapped,
copying a new one in could cause problems if it was already mapped.

The object files aren't content addressable and don't get mmapped,
so we need/can replace them with new ones (that might have new
content) during import.