Update of vector-algorithms library
Kioku has been updated to support the newest version
of the vector-algorithms library.

The biggest changes are that Lexicographic's terminate function
has been replaced by extent, and the size function's first
parameter is now of type Proxy, to signify the actual value
is meaningless, and all that matters is the type.

Note: The vector-algorithms library doesn't expose a terminate
function so we had to replicate it.