Fix "field" in parseTar to use len parameter
The "field" method that was defined in parseTar
didn't use the length parameter that was being
passed to it. This fix addresses that.

Further, this changes the name of "field" to
"tarField" so it doesn't shadow the "field"
method being imported from Memorizable.

Lastly, the "handler" parameter wasn't being used
by parseTar" in the first two definitions of it,
so that named paramter has been replaced with an
underscore.