Changes based on PR feedback
The cabal.config file has been removed as it
is no longer needed, since the project is now using
stack.

A constraint has been set on base, to be more clear
that we expect this library to work for a version of
base within the constraints given.

However, since this library has very limited usage,
we have done not due diligence to make sure it compiles
for all versions within the bounds given, since the
risk is low.