[ch14265] Possible fix for pending signals issue
It seems that the pending signals issue is related to
how Test.Tasty.HUnit throws exceptions when a test is
expected to pass, but fails instead.

Since we don't have an expectation that the result of the query
should be in the same order as the input, the tests have
been rewritten to use assertBool and a different way
to compare the lists has been implemented.

Additionally, code has been cleaned up to remove bits of cruft
that were added when tracing down errors.