Simplifies trieLookupMany by killing goKids
We observed that `goKids` was only being used when `remainingArc` was
empty.  Pulling at that thread led us down path of repeated
realizations that each step of the logic was actually meaningless,
culminating in the removal of `goKids` entirely.

Essentially, `breakCommonPrefixMultiKey` now contains the logic that
was originally supposed to be in `goKids` (but was incorrect in at
*least* the bug that was surfaced).

Co-Authored-By: Paul <paul@flipstone.com>