[ch15836] Addresses PR Feedback
- Minimizes the number of times that we deserialize trie nodes
- Makes `go` function tail-recursive
- Streamlines the code a bit

This works on the observation that we only need to read the trie root
elements when we have transitioned into one of it's subtries, if the
key has been completely consumed, or if a subtrie lookup returns
nothing.

it's fine to read the elems from a trie for which we don't know if the
arc is completely matched because `trieRootElems` only produces the
elems if the drop amount is equal to the length of it's arc.