Ignores cabal file from kioku build
Some time ago we removed the cabal file from Kioku but neglected to
ignore it.

While the cabal file is crucial to the build process itself it *is*
generated so it isn't necessary to check it in.

For libraries like kioku in particular it's annoying to have it checked
in because that means that as projects with varying versions of stack
(and thus hpack) build it then they will generate it slightly
differently leading to changing to the cabal file each time a different
project builds it.

For now I'm removing the file from the repo and ignoring it so that we
don't have unnecessary changes in the kioku submodule.