Avoid incomplete pattern match
Since we have `-Wall` and `-Werror` enabled, GHC 9.2 complains about
this being an incomplete pattern match.

This version of GHC cannot infer that the list can't be empty after
running 'sortBy' and 'nub' on a non-empty list. So the pattern match is
actually safe, but the compiler doesn't understand.

To avoid checking twice whether it is empty, we move the check until
after 'sortBy' and 'nub' have been run. Since we know that these
functions can work on empty lists, we now have equivalent behaviour,
even though it may be a tiny bit slower. This is a fatal error anyway,
so it shouldn't ever be called in a situation where performance matters.