Adds namespace support to Kioku
This feature establishes support for multiple subtrees of objects for a multi-tenant approach.
Now instead of the object directory housing the data set, index, and schema directories,
it holds namespace directories that have the data set, index and schema directories for those
particular namespaces.

This also cleans up some unused functions that were already broken and not used in any dependents we control.

co-authored-by: David <david@flipstone.com>